refactor: extract a character-like base from the character model
能力値・技能・派生値・技能判定の一式を CharacterLikeDataModel へ切り出す。
共鳴者（character）と、これから足す人間NPC（npc）が同じ判定計算を共有できるようにする。
共鳴値・共鳴感情・経歴は共鳴者だけの持ち物なので character 側に残す。挙動は不変。
